requirements.txt
flask==3.0.0
flask-cors==4.0.0
beautifulsoup4==4.12.2
requests==2.31.0
nltk==3.8.1
textstat==0.7.3
scikit-learn
numpy
lxml
python-dotenv==1.0.0
openai>=1.3.0
